Replace setup_intents class from active merchat v1.i98 with a decorator that will adapt the version in v1.119
